Improve the make-model-serial matching of outputs
Previously, the matching algorithm was looking for user-configured
output name as a substring in "$output_description ($output_name)". This
lead to unexpected behaviour when you have two outputs where the one's
description is a substring of the other's.

For example, say, you have monitors "hello (DP-1)" and "hello world
(DP-2)" and you're trying to match them in your Kanshi config using

profile {
  output "hello" …
  output "hello world" …
}

Substring matching allows "hello" to be matched with either of the
outputs, and the profile may end up being applied or rejected depending
on the order in which the outputs are discovered: if "hello world
(DP-2)" is discovered first, it gets matched with "hello", while "hello
world" fails to match with "hello (DP-1)".

This patch makes Kanshi discover the actual values of output's make,
model and serial; and then compares the user-supplied matching
string with "$make $model $serial" instead of looking for it in the
description. This is exactly how swaymsg forms the output descriptions
[1], so the user can simply copy-paste the output descriptions from
swaymsg into Kanshi config and have it uniquely identified.

To implement that, this patch takes advantage of the version 2 of
wlr-output-management-unstable-v1.xml protocol.

This patch fixes my issue [2] that I reported (I don't know if it fixes
the issues of others reported there).

[1]: https://github.com/swaywm/sway/blob/master/swaymsg/main.c#L218
[2]: https://todo.sr.ht/~emersion/kanshi/54