Pick lowest delta when fuzzy matching refreshes
If we have a refresh, try to pick something with an exact refresh first.
Doing a fuzzy match initially here can lead us to picking a refresh
such as 120.01 or 60.01, which is problematic for two reasons:
 - Modes such as 4K 120.01Hz is too much for link bandwidth of DP 1.4 without DSC.
 - It becomes out of phase with the majority of content being displayed.

Fixes me not being able to use 4K 120Hz on my monitor with AMDGPU + Radeon VII which supports DP 1.4, but not DSC (Display Stream Compression).

Signed-off-by: Joshua Ashton <joshua@froggi.es>